lambda-export/functions/telegram-bot-stack-TelegramBotFunction-fqeAhyC56WWc/requirements.txt
openai==1.51.2
requests==2.31.0
python-dotenv==1.0.0
google-cloud-bigquery==3.11.4
google-auth==2.23.0
google-api-core==2.11.1
httpx==0.27.0
boto3==1.34.0
boto3==1.34.0
